external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/rendering/point-size-expected.txt

Verify GL_VERTEX_PROGRAM_POINT_SIZE is enabled in WebGL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gl is non-null.
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
Draw a point of size 1 and verify it does not touch any other pixels.
PASS gl.getError() is gl.NO_ERROR
PASS buf[index] is 0
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 0
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 0
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 255
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
Draw a point of size 2 and verify it fills the appropriate region.
PASS gl.getError() is gl.NO_ERROR
PASS buf[index] is 255
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 255
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 255
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS buf[index] is 255
PASS buf[index + 1] is 0
PASS buf[index + 2] is 0
PASS successfullyParsed is true

TEST COMPLETE
